Wrapper Class
The wrapper class in Java provides the mechanism to convert primitive into object and object into primitive.

Change the value in Method: Java supports only call by value. So, if we pass a primitive value, it will not change the original value. But, if we convert the primitive value in an object, it will change the original value.